iot-test/raspberry-pi/requirements.txt
paho-mqtt
https://www.hivemq.com/demos/websocket-client/
